+++ tests/testthat/test-derive_param_tte.R
@@ -1,5 +1,10 @@
-test_that("new observations with analysis date are derived correctly", {
-  adsl <- tibble::tribble(
+library(tibble)
+library(lubridate)
+
+# derive_param_tte ----
+## Test 1: new observations with analysis date are derived correctly ----
+test_that("derive_param_tte Test 1: new observations with analysis date are derived correctly", {
+  adsl <-tribble(
     ~USUBJID, ~DTHFL, ~DTHDT,            ~LSTALVDT,         ~TRTSDT,           ~TRTSDTF,
     "03",     "Y",    ymd("2021-08-21"), ymd("2021-08-21"), ymd("2021-08-10"), NA,
     "04",     "N",    NA,                ymd("2021-05-24"), ymd("2021-02-03"), NA
@@ -28,7 +33,7 @@ test_that("new observations with analysis date are derived correctly", {
     )
   )
 
-  expected_output <- tibble::tribble(
+  expected_output <- tribble(
     ~USUBJID, ~ADT,              ~CNSR, ~EVENTDESC,              ~SRCDOM, ~SRCVAR,
     "03",     ymd("2021-08-21"), 0L,    "DEATH",                 "ADSL",  "DTHDT",
     "04",     ymd("2021-05-24"), 1L,    "LAST KNOWN ALIVE DATE", "ADSL",  "LSTALVDT"
@@ -59,7 +64,8 @@ test_that("new observations with analysis date are derived correctly", {
   )
 })
 
-test_that("new observations with analysis datetime are derived correctly", {
+## Test 2: new observations with analysis datetime are derived correctly ----
+test_that("derive_param_tte Test 2: new observations with analysis datetime are derived correctly", {
   adsl <- tibble::tribble(
     ~USUBJID, ~DTHFL, ~DTHDT,            ~TRTSDTM,                       ~TRTSDTF, ~TRTSTMF,
     "01",     "Y",    ymd("2021-06-12"), ymd_hms("2021-01-01 00:00:00"), "M",      "H",
@@ -128,7 +134,7 @@ test_that("new observations with analysis datetime are derived correctly", {
   )
 
   # nolint start
-  expected_output <- tibble::tribble(
+  expected_output <- tribble(
     ~USUBJID, ~ADTM,                          ~CNSR, ~EVENTDESC,              ~SRCDOM, ~SRCVAR,   ~SRCSEQ,
     "01",     ymd_hms("2021-05-05 12:02:00"), 0L,    "PD",                    "ADRS",  "ADTM",    3,
     "02",     ymd_hms("2021-02-03 10:56:00"), 0L,    "PD",                    "ADRS",  "ADTM",    1,
@@ -167,7 +173,8 @@ test_that("new observations with analysis datetime are derived correctly", {
   )
 })
 
-test_that("new observations based on DTC variables are derived correctly", {
+## Test 3: error is issued if DTC variables specified for date ----
+test_that("derive_param_tte Test 3: error is issued if DTC variables specified for date", {
   adsl <- tibble::tribble(
     ~USUBJID, ~TRTSDT,           ~EOSDT,
     "01",     ymd("2020-12-06"), ymd("2021-03-06"),
@@ -205,19 +212,7 @@ test_that("new observations based on DTC variables are derived correctly", {
     )
   )
 
-  expected_output <- tibble::tribble(
-    ~USUBJID, ~ADT,              ~CNSR, ~EVENTDESC,     ~SRCDOM, ~SRCVAR,   ~SRCSEQ,
-    "01",     ymd("2021-01-01"), 0L,    "AE",           "AE",    "AESTDTC", 3,
-    "02",     ymd("2021-02-03"), 1L,    "END OF STUDY", "ADSL",  "EOSDT",   NA
-  ) %>%
-    mutate(
-      STUDYID = "AB42",
-      PARAMCD = "TTAE",
-      PARAM = "Time to First Adverse Event"
-    ) %>%
-    left_join(select(adsl, USUBJID, STARTDT = TRTSDT), by = "USUBJID")
-
-  expect_dfs_equal(
+  expect_error(
     derive_param_tte(
       dataset_adsl = adsl,
       start_date = TRTSDT,
@@ -229,12 +224,12 @@ test_that("new observations based on DTC variables are derived correctly", {
         PARAM = "Time to First Adverse Event"
       )
     ),
-    expected_output,
-    keys = c("USUBJID", "PARAMCD")
+    regexp = "AESTDTC specified for `date` is not a date or datetime variable but is a string"
   )
 })
 
-test_that("by_vars parameter works correctly", {
+## Test 4: by_vars parameter works correctly ----
+test_that("derive_param_tte Test 4: by_vars parameter works correctly", {
   adsl <- tibble::tribble(
     ~USUBJID, ~TRTSDT,           ~EOSDT,
     "01",     ymd("2020-12-06"), ymd("2021-03-06"),
@@ -243,16 +238,17 @@ test_that("by_vars parameter works correctly", {
     mutate(STUDYID = "AB42")
 
   ae <- tibble::tribble(
-    ~USUBJID, ~AESTDTC,           ~AESEQ, ~AEDECOD,
-    "01",     "2021-01-03T10:56", 1,      "Flu",
-    "01",     "2021-03-04",       2,      "Cough",
-    "01",     "2021",             3,      "Flu"
+    ~USUBJID, ~AESTDTC,     ~AESEQ, ~AEDECOD,
+    "01",     "2021-01-03", 1,      "Flu",
+    "01",     "2021-03-04", 2,      "Cough",
+    "01",     "2021-01-01", 3,      "Flu"
   ) %>%
-    mutate(STUDYID = "AB42")
+    mutate(STUDYID = "AB42",
+           AESTDT = ymd(AESTDTC))
 
   ttae <- event_source(
     dataset_name = "ae",
-    date = AESTDTC,
+    date = AESTDT,
     set_values_to = vars(
       EVENTDESC = "AE",
       SRCDOM = "AE",
@@ -308,25 +304,27 @@ test_that("by_vars parameter works correctly", {
   )
 })
 
-test_that("an error is issued if some of the by variables are missing", {
-  adsl <- tibble::tribble(
+## Test 5: an error is issued if some of the by variables are missing ----
+test_that("derive_param_tte Test 5: an error is issued if some of the by variables are missing", {
+  adsl <- tribble(
     ~USUBJID, ~TRTSDT,           ~EOSDT,
     "01",     ymd("2020-12-06"), ymd("2021-03-06"),
     "02",     ymd("2021-01-16"), ymd("2021-02-03")
   ) %>%
     mutate(STUDYID = "AB42")
 
-  ae <- tibble::tribble(
-    ~USUBJID, ~AESTDTC,           ~AESEQ, ~AEDECOD,
-    "01",     "2021-01-03T10:56", 1,      "Flu",
-    "01",     "2021-03-04",       2,      "Cough",
-    "01",     "2021",             3,      "Flu"
+  ae <- tribble(
+    ~USUBJID, ~AESTDTC,     ~AESEQ, ~AEDECOD,
+    "01",     "2021-01-03", 1,      "Flu",
+    "01",     "2021-03-04", 2,      "Cough",
+    "01",     "2021-01-01", 3,      "Flu"
   ) %>%
-    mutate(STUDYID = "AB42")
+    mutate(STUDYID = "AB42",
+           AESTDT = ymd(AESTDTC))
 
   ttae <- event_source(
     dataset_name = "ae",
-    date = AESTDTC,
+    date = AESTDT,
     set_values_to = vars(
       EVENTDESC = "AE",
       SRCDOM = "AE",
@@ -364,8 +362,10 @@ test_that("an error is issued if some of the by variables are missing", {
     regexp = "^Only AEDECOD are included in source dataset.*"
   )
 })
-test_that("an error is issued all by variables are missing in all source datasets", {
-  adsl <- tibble::tribble(
+
+## Test 6: an error is issued all by variables are missing in all source datasets ----
+test_that("derive_param_tte Test 6: an error is issued if all by variables are missing in all source datasets", {
+  adsl <- tribble(
     ~USUBJID, ~TRTSDT,           ~EOSDT,
     "01",     ymd("2020-12-06"), ymd("2021-03-06"),
     "02",     ymd("2021-01-16"), ymd("2021-02-03")
@@ -373,16 +373,17 @@ test_that("an error is issued all by variables are missing in all source dataset
     mutate(STUDYID = "AB42")
 
   ae <- tibble::tribble(
-    ~USUBJID, ~AESTDTC,           ~AESEQ, ~AEDECOD,
-    "01",     "2021-01-03T10:56", 1,      "Flu",
-    "01",     "2021-03-04",       2,      "Cough",
-    "01",     "2021",             3,      "Flu"
+    ~USUBJID, ~AESTDTC,     ~AESEQ, ~AEDECOD,
+    "01",     "2021-01-03", 1,      "Flu",
+    "01",     "2021-03-04", 2,      "Cough",
+    "01",     "2021-01-01", 3,      "Flu"
   ) %>%
-    mutate(STUDYID = "AB42")
+    mutate(STUDYID = "AB42",
+           AESTDT = ymd(AESTDTC))
 
   ttae <- event_source(
     dataset_name = "ae",
-    date = AESTDTC,
+    date = AESTDT,
     set_values_to = vars(
       EVENTDESC = "AE",
       SRCDOM = "AE",
@@ -422,25 +423,27 @@ test_that("an error is issued all by variables are missing in all source dataset
   )
 })
 
-test_that("an error is issued if there is no one to one mapping between PARAMCD and by_vars", {
-  adsl <- tibble::tribble(
+## Test 7: an error is issued if there is no one to one mapping between PARAMCD and by_vars ----
+test_that("derive_param_tte Test 7: an error is issued if there is no one to one mapping between PARAMCD and by_vars", {
+  adsl <- tribble(
     ~USUBJID, ~TRTSDT,           ~EOSDT,
     "01",     ymd("2020-12-06"), ymd("2021-03-06"),
     "02",     ymd("2021-01-16"), ymd("2021-02-03")
   ) %>%
     mutate(STUDYID = "AB42")
 
-  ae <- tibble::tribble(
-    ~USUBJID, ~AESTDTC,           ~AESEQ, ~AEDECOD,
-    "01",     "2021-01-03T10:56", 1,      "Flu",
-    "01",     "2021-03-04",       2,      "Cough",
-    "01",     "2021",             3,      "Flu"
+  ae <- tribble(
+    ~USUBJID, ~AESTDTC,     ~AESEQ, ~AEDECOD,
+    "01",     "2021-01-03", 1,      "Flu",
+    "01",     "2021-03-04", 2,      "Cough",
+    "01",     "2021-01-01", 3,      "Flu"
   ) %>%
-    mutate(STUDYID = "AB42")
+    mutate(STUDYID = "AB42",
+           AESTDT = ymd(AESTDTC))
 
   ttae <- event_source(
     dataset_name = "ae",
-    date = AESTDTC,
+    date = AESTDT,
     set_values_to = vars(
       EVENTDESC = "AE",
       SRCDOM = "AE",
@@ -481,25 +484,27 @@ test_that("an error is issued if there is no one to one mapping between PARAMCD 
   )
 })
 
-test_that("an error if issued set_values_to contains invalid expressions", {
-  adsl <- tibble::tribble(
+## Test 8: an error if issued set_values_to contains invalid expressions ----
+test_that("derive_param_tte Test 8: an error if issued set_values_to contains invalid expressions", {
+  adsl <- tribble(
     ~USUBJID, ~TRTSDT,           ~EOSDT,
     "01",     ymd("2020-12-06"), ymd("2021-03-06"),
     "02",     ymd("2021-01-16"), ymd("2021-02-03")
   ) %>%
     mutate(STUDYID = "AB42")
 
-  ae <- tibble::tribble(
-    ~USUBJID, ~AESTDTC,           ~AESEQ, ~AEDECOD,
-    "01",     "2021-01-03T10:56", 1,      "Flu",
-    "01",     "2021-03-04",       2,      "Cough",
-    "01",     "2021",             3,      "Flu"
+  ae <- tribble(
+    ~USUBJID, ~AESTDTC,     ~AESEQ, ~AEDECOD,
+    "01",     "2021-01-03", 1,      "Flu",
+    "01",     "2021-03-04", 2,      "Cough",
+    "01",     "2021-01-01", 3,      "Flu"
   ) %>%
-    mutate(STUDYID = "AB42")
+    mutate(STUDYID = "AB42",
+           AESTDT = ymd(AESTDTC))
 
   ttae <- event_source(
     dataset_name = "ae",
-    date = AESTDTC,
+    date = AESTDT,
     set_values_to = vars(
       EVENTDESC = "AE",
       SRCDOM = "AE",
@@ -548,95 +553,27 @@ test_that("an error if issued set_values_to contains invalid expressions", {
   )
 })
 
-test_that("new observations analysis datetime based on DTC variables are derived correctly", {
-  adsl <- tibble::tribble(
-    ~USUBJID, ~TRTSDTM,                       ~EOSDT,
-    "01",     ymd_hms("2020-12-06 14:23:00"), ymd("2021-03-06"),
-    "02",     ymd_hms("2021-01-16 13:09:00"), ymd("2021-02-03")
-  ) %>%
-    mutate(STUDYID = "AB42")
-
-  ae <- tibble::tribble(
-    ~USUBJID, ~AESTDTC,           ~AESEQ,
-    "01",     "2021-01-03T10:56", 1,
-    "01",     "2021-03-04",       2,
-    "01",     "2021-03",          3
-  ) %>%
-    mutate(STUDYID = "AB42")
-
-  ttae <- event_source(
-    dataset_name = "ae",
-    date = AESTDTC,
-    set_values_to = vars(
-      EVENTDESC = "AE",
-      SRCDOM = "AE",
-      SRCVAR = "AESTDTC",
-      SRCSEQ = AESEQ
-    )
-  )
-
-  eos <- censor_source(
-    dataset_name = "adsl",
-    date = EOSDT,
-    censor = 1,
-    set_values_to = vars(
-      EVENTDESC = "END OF STUDY",
-      SRCDOM = "ADSL",
-      SRCVAR = "EOSDT"
-    )
-  )
-
-  expected_output <- tibble::tribble(
-    ~USUBJID, ~ADTM,                          ~CNSR, ~EVENTDESC,     ~SRCDOM, ~SRCVAR,   ~SRCSEQ,
-    "01",     ymd_hms("2021-01-03 10:56:00"), 0L,    "AE",           "AE",    "AESTDTC", 1,
-    "02",     ymd_hms("2021-02-03 00:00:00"), 1L,    "END OF STUDY", "ADSL",  "EOSDT",   NA
-  ) %>%
-    mutate(
-      STUDYID = "AB42",
-      PARAMCD = "TTAE",
-      PARAM = "Time to First Adverse Event"
-    ) %>%
-    left_join(select(adsl, USUBJID, STARTDTM = TRTSDTM), by = "USUBJID")
-
-  actual_output <- derive_param_tte(
-    dataset_adsl = adsl,
-    start_date = TRTSDTM,
-    event_conditions = list(ttae),
-    censor_conditions = list(eos),
-    source_datasets = list(adsl = adsl, ae = ae),
-    create_datetime = TRUE,
-    set_values_to = vars(
-      PARAMCD = "TTAE",
-      PARAM = "Time to First Adverse Event"
-    )
-  )
-
-  expect_dfs_equal(
-    actual_output,
-    expected_output,
-    keys = c("USUBJID", "PARAMCD")
-  )
-})
-
-test_that("error is issued if parameter code already exists", {
-  adsl <- tibble::tribble(
+## Test 9: error is issued if parameter code already exists ----
+test_that("derive_param_tte Test 9: error is issued if parameter code already exists", {
+  adsl <- tribble(
     ~USUBJID, ~TRTSDT,           ~EOSDT,
     "01",     ymd("2020-12-06"), ymd("2021-03-06"),
     "02",     ymd("2021-01-16"), ymd("2021-02-03")
   ) %>%
     mutate(STUDYID = "AB42")
 
-  ae <- tibble::tribble(
-    ~USUBJID, ~AESTDTC,           ~AESEQ,
-    "01",     "2021-01-03T10:56", 1,
-    "01",     "2021-03-04",       2,
-    "01",     "2021",             3
+  ae <- tribble(
+    ~USUBJID, ~AESTDTC,     ~AESEQ, ~AEDECOD,
+    "01",     "2021-01-03", 1,      "Flu",
+    "01",     "2021-03-04", 2,      "Cough",
+    "01",     "2021-01-01", 3,      "Flu"
   ) %>%
-    mutate(STUDYID = "AB42")
+    mutate(STUDYID = "AB42",
+           AESTDT = ymd(AESTDTC))
 
   ttae <- event_source(
     dataset_name = "ae",
-    date = AESTDTC,
+    date = AESTDT,
     set_values_to = vars(
       EVENTDESC = "AE",
       SRCDOM = "AE",
@@ -685,7 +622,9 @@ test_that("error is issued if parameter code already exists", {
   )
 })
 
-test_that("`tte_source` objects are printed as intended", {
+# print.tte_source ----
+## Test 10: tte_source` objects are printed as intended ----
+test_that("`print.tte_source Test 10: tte_source` objects are printed as intended", {
   ttae <- event_source(
     dataset_name = "ae",
     date = AESTDTC,